\name{awesomelab3-package}
\alias{awesomelab3-package}
\alias{awesomelab3}
\docType{package}
\title{
\packageTitle{awesomelab3}
}
\description{
\packageDescription{awesomelab3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{awesomelab3}
\packageIndices{awesomelab3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{awesomelab3}

Maintainer: \packageMaintainer{awesomelab3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
