\name{awesomelab3-package}
\alias{awesomelab3-package}
\alias{awesomelab3}
\docType{package}
\title{
\packageTitle{awesomelab3}
}
\description{
\packageDescription{awesomelab3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{awesomelab3}
\packageIndices{awesomelab3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{awesomelab3}

Maintainer: \packageMaintainer{awesomelab3}
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\examples{
euclidean(10, 100)
}
