% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{awesomelab3}
\alias{awesomelab3}
\alias{awesomelab3-package}
\title{awesomelab3: learning to create a package in R.}
\description{
The awesomelab3 package provides two diferent and very famous algorithms:
Euclidean algorithm and Dijkstra's algorithm.
}
\section{Functions}{

\code{euclidean()} returns the greatest common divisor between scalars a and b.

\code{dijkstra()} returns the shortest path in a graph.
}

\author{
Martin Svensson and Laura Julia Melis.

Course in Advanced Programming in R.

Departmenf of Computer and Information Science, Linköping University.
}
