% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{dijkstra Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{??}

\item{init_node}{scalar}
}
\value{
path?
}
\description{
#TODO#
}
\examples{
data(wiki_graph)
dijkstra(wiki_graph, 1)

}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
