% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{scalar}

\item{b}{scalar}
}
\value{
Scalar value of the greatest common divisor.
}
\description{
Euclidean algorithm finds the greatest common divisor between scalars \code{a} and \code{b}.
}
\examples{
euclidean(a, b)

}
