% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Graph of nodes and distances between nodes.}
\format{A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{a node of the graph}
  \item{v2}{a node of the graph, adjacent node of v1}
  \item{w}{the distance of the path between v1 and v2}
}}
\source{
\url{https://en.wikipedia.org/wiki/Graph_theory}
}
\usage{
wiki_graph
}
\description{
A dataset containing the nodes of a graph and the distances between them
(it is the first graph at the wikipedia page, see References).
}
\keyword{datasets}
