% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{numeric scalar object.}

\item{b}{numeric scalar object.}
}
\value{
\code{euclidean} returns a scalar value of the greatest common divisor.

If the arguments of the function are not numeric scalars or integers an error will occur.
}
\description{
Euclidean algorithm finds the greatest common divisor (gcd) between scalars \code{a} and \code{b}.
}
\details{
The algorithm works by following two rules.

(1) If \code{b=0}, then \code{gcd(a,b)=a} and the algorithm ends.

(2) Elsewhere (\code{b!=0}), then \code{gcd(a,b)=gcd(b,r)} where \code{r} is the reminder of dividing \code{a} by \code{b}. \code{gcd(a,b)=gcd(b,r)} will be calculated until the remainder equals 0.
}
\examples{
euclidean(123612, 13892347912)  # = 4
euclidean(10, 100)              # = 10

}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
